HSEARCH-2689 Retrieve the EntityManagerFactoryRegistry through injection only
... instead of relying on the JobContextSetupListener being overridden.

This will allow to move the job context setup code to a static util
class, which in turn will allow to setup the job context from multiple
classes.

This is necessary because the spec actually states that the job
context is not shared between threads (see
https://github.com/WASdev/standards.jsr352.jbatch/issues/50), meaning we
have to setup the context for each new partition.