HSEARCH-2689 Do not use the Session in LuceneDocProducer
Passing information from one batch job component to another seems more
and more likely to be painful, so I'd rather do it only when strictly
necessary. On top of that, the current use of the session context is
dubious at best, since we put two different types of objects in
stepContext.getTransientUserData (StepProgress and
PartitionContextData). I am trying to untangle this.

Besides, we should not need the session in LuceneDocProducer, as entity
loading is supposed to be the EntityReader's responsibility.

This change required to:

 * get the SearchIntegrator from something else than the session
(namely, from the EntityManagerFactory).
 * not use HibernateSessionLoadingInitializer, which requires a specific
session, but rather the default initializer.

Luckily, both changes made sense regardless of the purpose of this
commit: that's probably how the doc producer should have been in the
first place.