HSEARCH-2689 Fix the scope of batch artifacts
We used to assign a Singleton scope, which is utterly wrong: those
artifacts should be instantiated at least for each job, and some should
even be instantiated for each partition.

As it happens, the "job" and "partition" scope are not defined in the
JSR-352 spec, but only in implementations... As a result:

 * we have to define our own scope annotations
 * we have to ask to integrators to map those scope annotations to
whatever scope is defined by their JSR-352 implementations.
 * we cannot have a generic "CDI" integration anymore, but must target
the more specific "JBeret" instead.